requirements.txt
Django==4.2.7
PyMySQL
python-dotenv==1.0.0
Pillow>=11.0.0
